# apps/web/openapi/shipgate.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Shipgate API",
    "description": "Shipgate HTTP API.",
    "version": "0.0.0"
  },
  "components": {
    "securitySchemes": {
      "shipgateSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-shipgate_session",
        "description": "HttpOnly Shipgate browser session cookie."
      }
    },
...
  },
  "paths": {
...
    "/api/v1/projects/{projectId}/reconciliation": {
      "post": {
        "operationId": "reconcileProject",
        "summary": "Queue an authoritative repository reconciliation",
        "tags": ["Projects"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["expectedConfigurationVersion"],
                "properties": {
                  "expectedConfigurationVersion": {
                    "type": "integer",
                    "minimum": 1
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "in": "path",
            "name": "projectId",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 256
            },
            "in": "header",
            "name": "x-csrf-token",
            "required": false
          }
        ],
        "security": [
          {
            "shipgateSession": []
          }
        ],
        "responses": {
          "202": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["reconciliation"],
                  "properties": {
                    "reconciliation": {
                      "type": "object",
                      "required": [
                        "requestId",
                        "status",
                        "configurationVersion",
                        "reason",
                        "mode",
                        "sourceSha",
                        "productionSha",
                        "requestedAt"
                      ],
                      "properties": {
                        "requestId": {
                          "type": "string"
                        },
                        "status": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": ["queued"]
                            },
                            {
                              "type": "string",
                              "enum": ["running"]
                            },
                            {
                              "type": "string",
                              "enum": ["succeeded"]
                            },
                            {
                              "type": "string",
                              "enum": ["superseded"]
                            },
                            {
                              "type": "string",
                              "enum": ["failed"]
                            },
                            {
                              "type": "string",
                              "enum": ["cancelled"]
                            }
                          ]
                        },
                        "configurationVersion": {
                          "type": "integer",
                          "minimum": 1
                        },
                        "reason": {
                          "type": "string"
                        },
                        "mode": {
                          "type": "string",
                          "enum": ["full"]
                        },
                        "sourceSha": {
                          "type": "string"
                        },
                        "productionSha": {
                          "type": "string"
                        },
                        "requestedAt": {
                          "type": "string"
                        }
                      },
                      "additionalProperties": false
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Common error response returned by the Shipgate API.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["code", "message", "requestId"],
                  "properties": {
                    "code": {
                      "type": "string",
                      "minLength": 1,
                      "example": "VALIDATION_ERROR"
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1
                    },
                    "requestId": {
                      "type": "string",
                      "minLength": 1
                    },
                    "details": {}
                  },
                  "additionalProperties": false,
                  "title": "ApiError",
                  "description": "Common error response returned by the Shipgate API."
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Shipgate origin"
    }
  ],
...
}